refactor(data-analysis): align module with frontend code style
