+++ yak-ops-ui/src/pages/data-analysis/lineage/components/service.ts
@@ -1,0 +1,5 @@
+export {
+  getLineageAssetByKey as fetchLineageAssetByKey,
+  getLineageGraph as fetchLineageGraph,
+  searchLineageAssets,
+} from '@/services/data-analysis';
