+++ yak-ops-ui/src/pages/data-analysis/lineage/components/types.ts
@@ -1,0 +1,2 @@
+export { assetTypeLabel } from '@/services/data-analysis';
+export type { LineageAsset, LineageAssetType } from '@/services/data-analysis';
