+++ yak-ops-ui/src/pages/data-analysis/lineage/utils/types.ts
@@ -1,0 +1,8 @@
+export type {
+  LineageAsset,
+  LineageAssetType,
+  LineageDirection,
+  LineageGraph,
+  LineageRelation,
+  LineageRelationType,
+} from '@/services/data-analysis';
