+++ yak-ops-ui/src/services/data-analysis/index.ts
@@ -1,0 +1,3 @@
+export * from './catalog';
+export * from './lineage';
+export * from './types';
